Add files via upload
This project aims to develop a real-time sentiment analysis system for phone calls. The system leverages Python's SpeechRecognizer library to convert spoken language into text and uses TextBlob for sentiment analysis. The goal is to determine the sentiment of a phone call—whether it is positive, negative, or neutral—in real-time.